move LoginHelper in Auth namespace
