+++ src/Controllers/LoginController.php
@@ -22,6 +22,7 @@ use Elabftw\Auth\Local;
 use OneLogin\Saml2\Response as SamlResponse;
 use OneLogin\Saml2\Settings as SamlSettings;
 use Elabftw\Auth\LoginFlow;
+use Elabftw\Auth\LoginHelper;
 use Elabftw\Auth\MfaRateLimiter;
 use Elabftw\Auth\MfaRequired;
 use Elabftw\Auth\PasswordRenewalRequired;
@@ -56,7 +57,6 @@ use Elabftw\Models\Users2Teams;
 use Elabftw\Services\DeviceToken;
 use Elabftw\Services\DeviceTokenValidator;
 use Elabftw\Services\Filter;
-use Elabftw\Services\LoginHelper;
 use Elabftw\Services\TeamsHelper;
 use Elabftw\Services\ResetPasswordKey;
 use LdapRecord\Connection;
