+++ web/app/init.inc.php
@@ -17,9 +17,9 @@ use Elabftw\Auth\AnonymousLoginValidator;
 use Elabftw\Auth\Cookie;
 use Elabftw\Auth\CookieLogin;
 use Elabftw\Auth\CookieToken;
+use Elabftw\Auth\LoginHelper;
 use Elabftw\Auth\UserLoginValidator;
 use Elabftw\Enums\EntityType;
-use Elabftw\Enums\Entrypoint;
 use Elabftw\Enums\Language;
 use Elabftw\Exceptions\DatabaseErrorException;
 use Elabftw\Exceptions\ImproperActionException;
@@ -27,7 +27,6 @@ use Elabftw\Exceptions\UnauthorizedException;
 use Elabftw\Models\Config;
 use Elabftw\Models\Users\AnonymousUser;
 use Elabftw\Models\Users\Users;
-use Elabftw\Services\LoginHelper;
 use Elabftw\Services\TeamFinder;
 use Exception;
 use PDOException;
